requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# AWS and GitHub dependencies
boto3>=1.26.0
botocore>=1.29.0
PyGithub>=2.0.0

# Strands Agents SDK
strands-agents>=0.1.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
